LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-new-navigation-before-commit-expected.txt


FAIL Cancel a navigation with a precommitHandler before commit by starting a new navigation assert_equals: expected "" but got "#ShouldNotCommit"
